dsbd_project/data_storage/requirements.txt
confluent_kafka
mysql.connector